requirements.txt
alembic==1.13.2
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7

dnspython==2.6.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
# is-disposable-email==1.0.0
# itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.0
pillow==10.4.0
pluggy==1.4.0
pytest==8.1.1
pywin32==306
requests==2.31.0
SQLAlchemy==2.0.30
# typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2
wtforms-validators==1.0.0
